new file:   scripts/launchers/Caddyfile_darwin; modified:   scripts/launchers/various-darwin.zsh; new file:   scripts/launchers/wsgidav_darwin.yaml; modified:   scripts/setup/brewables; modified:   scripts/zshlang/auto-load/others/cleaners.zsh; modified:   scripts/zshlang/auto-load/others/filesystem.zsh; modified:   scripts/zshlang/auto-load/others/network.zsh; modified:   scripts/zshlang/auto-load/others/telegram.zsh; modified:   scripts/zshlang/basic/auto-load/enhanced.zsh; modified:   scripts/zshlang/basic/auto-load/filesystem.zsh; modified:   scripts/zshlang/basic/auto-load/path.zsh; modified:   scripts/zshlang/interactive/auto-load/fuzzy/fuzzy.zsh; modified:   scripts/zshlang/interactive/auto-load/fuzzy/notes-search.zsh; modified:   .bashrc

+++ scripts/zshlang/auto-load/others/cleaners.zsh
@@ -5,18 +5,30 @@ rm-deleteus() {
 }
 ##
 function rm-caches-sudo() {
-    if isLinux ; then
+    if isLinux && ! isBorg ; then
         sudo apt-get clean
         sudo apt-get autoclean
         sudo apt-get autoremove -y
         # sudo rm -rf /var/log/* # this can break things that depend on some directories existing. Delete manually via ncdu.
     fi
 }
+function trs-rm() {
+    local paths=("$@")
+
+    local p
+    for p in $paths[@] ; do
+        if test -e "$p" ; then
+            reval-ec command rm -rf "$p"
+        fi
+    done
+}
 function rm-caches() {
     rm-deleteus
-    silent command rm -r ~/.local/share/Trash/*(DN)
-    silent command rm ~/sent
-    command rm -rf "$(brew --cache)"
+    trs-rm  ~/.local/share/Trash/*(DN)
+    trs-rm ~/sent
+    trs-rm ~/logs ~/log
+    trs-rm ~/julia_tmp/ ~/tmp-kindle
+    trs-rm "$(brew --cache)"
     conda clean --all --yes # STILL does not delete old python version stuff!
     trs-empty
 }
@@ -24,4 +36,5 @@ function rm-caches-all() {
     rm-caches-sudo
     rm-caches
 }
+aliasfn cleanup rm-caches-all
 ##
